+++ Facturacion/backend/modelos/presupuesto_bd.py
@@ -11,7 +11,7 @@ class PresupuestoBd(BaseBd):
     valido_hasta= Column(Date, nullable=False)
     tipo_comprobante= Column(String, nullable=False)
     
-    cuit_vendedor= Column(Integer, ForeignKey("vendedores.cuit"))
+    cuit_vendedor= Column(BIGINT, ForeignKey("vendedores.cuit"))
 
     cuit_cliente= Column(BIGINT, ForeignKey("clientes.cuit"), nullable=False)
